% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaisi.R
\name{configure_daisi}
\alias{configure_daisi}
\title{Set the necessary environment variables to configure the Daisi client}
\usage{
configure_daisi(python_path = NULL, daisi_instance = "app")
}
\arguments{
\item{python_path}{The path to the python binary on your system}

\item{daisi_instance}{The Daisi instance to connect to}
}
\value{
TRUE if successful update has occurred
}
\description{
Set the necessary environment variables to configure the Daisi client
}
\examples{
\dontrun{
configure_daisi(python_path = "/usr/local/bin/python3", daisi_instance = "app")
}
}
