% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaisi.R
\name{Daisi}
\alias{Daisi}
\title{Initialize and connect to the given Daisi}
\usage{
Daisi(daisi_id, base_url = NULL)
}
\arguments{
\item{daisi_id}{The name or UUID of the daisi}

\item{base_url}{The platform on which to access the daisi}
}
\value{
daisi object with daisi information
}
\description{
Initialize and connect to the given Daisi
}
\examples{
\dontrun{
configure_daisi()

d <- Daisi("Add Two Numbers")
d
}
}
