% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaisi.R
\name{DaisiExecution}
\alias{DaisiExecution}
\title{Generate a new execution of a given Daisi}
\usage{
DaisiExecution(daisi, endpoint, ...)
}
\arguments{
\item{daisi}{The Daisi object, initialized with Daisi()}

\item{endpoint}{The endpoint of the Daisi to call}

\item{...}{Arguments passed onto the underlying Daisi}
}
\value{
DaisiExecution object with the Execution parameters
}
\description{
Generate a new execution of a given Daisi
}
\examples{
\dontrun{
configure_daisi()

d <- Daisi("Add Two Numbers")

de <- DaisiExecution(d, list(firstNumber = 5, secondNumber = 6))

Sys.sleep(1)

de$value()
}

}
