% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaisi.R
\name{DaisiMapExecution}
\alias{DaisiMapExecution}
\title{Generate a new map execution of a given Daisi}
\usage{
DaisiMapExecution(daisi, args_list)
}
\arguments{
\item{daisi}{The Daisi object, initialized with Daisi()}

\item{args_list}{A list of named lists of arguments to provide to the Daisi}
}
\value{
DaisiExecution object with the Execution parameters
}
\description{
Generate a new map execution of a given Daisi
}
\examples{
\dontrun{
configure_daisi()

d <- Daisi("Add Two Numbers")

bulk_options <- lapply(1:10, function(x) {
    list(firstNumber = x, secondNumber = 5)
})

deb <- DaisiMapExecution(d, bulk_options)
deb
}
}
