% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaisi.R
\name{execute_daisi}
\alias{execute_daisi}
\title{Begin execution of the given Daisi}
\usage{
execute_daisi(daisi_execution)
}
\arguments{
\item{daisi_execution}{The DaisiExecution object for which to fetch the result}
}
\value{
The ID (in UUID format) of the execution
}
\description{
Begin execution of the given Daisi
}
