% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaisi.R
\name{result_daisi}
\alias{result_daisi}
\title{Get the result of the Daisi Execution}
\usage{
result_daisi(daisi_execution)
}
\arguments{
\item{daisi_execution}{The DaisiExecution object for which to fetch the result}
}
\value{
The object produced by the given Daisi
}
\description{
Get the result of the Daisi Execution
}
